passing test for POST /lists route
