+++ lib/app.ts
@@ -1,5 +1,6 @@
 import express from 'express';
 import usersRouter from './routes/users.js';
+import listsRouter from './routes/lists.js';
 import notFoundMiddleWare from './middleware/not-found.js';
 import errorMiddleware from './middleware/error.js';
 import cors from 'cors';
@@ -14,6 +15,7 @@ app.use(cors());
 
 // App routes
 app.use('/users', usersRouter);
+app.use('/lists', listsRouter);
 
 // Error handling & 404 middleware for when
 // a request doesn't match any app routes
