+++ tests/utils.ts
@@ -1,5 +1,5 @@
 import { readFileSync } from 'node:fs';
-import pool from '../sql/pool';
+import pool from '../sql/pool.js';
 const sql = readFileSync('./sql/setup.sql', 'utf-8');
 
 function setupDb() {
